requirements.txt
Flask
flask-restx
SQLAlchemy
psycopg2-binary
python-dotenv
PyYAML
gunicorn
docker
